requirements.txt
flask
python-dotenv
yfinance
pandas
numpy
requests
gunicorn
apscheduler
lxml
